[PIWEB-20373] refactor: Restructure automation context to match the common structure
